Refactor view.php into modular includes
